+++ packages/auth/src/Domain/Token/ValetTokenData.ts
@@ -1,0 +1,9 @@
+import { Uuid } from '@standardnotes/common'
+
+export type ValetTokenData = {
+  userUuid: Uuid,
+  permittedOperation: 'read' | 'write',
+  permittedResources: Array<string>,
+  uploadBytesUsed: number,
+  uploadBytesLimit: number,
+}