+++ packages/snjs/lib/services/component_manager/component_viewer.ts
@@ -468,10 +468,7 @@ export class ComponentViewer {
     return responseItems.map((responseItem) => {
       const privateProperties = privateContentProperties.slice();
       /** Server extensions are allowed to modify url property */
-      if (
-        removeUrls &&
-        responseItem.content_type !== ContentType.ServerExtension
-      ) {
+      if (removeUrls) {
         privateProperties.push('url');
       }
       if (!responseItem.content || isString(responseItem.content)) {
@@ -766,21 +763,15 @@ export class ComponentViewer {
           this.component.uuid
         );
         this.syncService
-          .sync()
-          .then(() => {
-            /* Allow handlers to be notified when a save begins and ends, to update the UI */
-            const saveMessage = Object.assign({}, message);
-            saveMessage.action = ComponentAction.SaveSuccess;
-            this.replyToMessage(message, {});
-            this.handleMessage(saveMessage);
+          .sync({
+            onPresyncSave: () => {
+              this.replyToMessage(message, {});
+            },
           })
           .catch(() => {
-            const saveMessage = Object.assign({}, message);
-            saveMessage.action = ComponentAction.SaveError;
             this.replyToMessage(message, {
-              error: ComponentAction.SaveError,
+              error: 'save-error',
             });
-            this.handleMessage(saveMessage);
           });
       }
     );